// admin/src/api/generated/schemas/adminAdminPermission.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

/**
 * AdminPermission represents a permission (admin_permission).
 */
export type AdminAdminPermission = {
  id?: string;
  name?: string;
  code?: string;
  description?: string;
  menu_id?: string;
  status?: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
};