// admin/src/api/generated/schemas/adminCreatePermissionRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminCreatePermissionRequest = {
  name?: string;
  code?: string;
  menu_id?: string;
  description?: string;
  status?: number;
};